% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage}
\format{
A data frame with 15 rows and 2 columns:
\describe{
  \item{DISTANCE}{Distance from the fire station in miles}
  \item{DAMAGE}{Dollar cost of damage in thousands}
  ...
}
}
\source{
<https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sinich/p/book/9781498728850>
}
\usage{
fire
}
\description{
A dataset containing the values for fire damage at accident locations.
}
\keyword{datasets}
