% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{My Central Limit Theorem Function}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{Number of uniform random variables to sum in each iteration}

\item{iter}{Number of iterations (sample size for the distribution of the sum of uniform random variables)}

\item{a}{Lower bound of the uniform distribution}

\item{b}{Upper bound of the uniform distribution}
}
\value{
A histogram of the distribution of the sum of uniform random variables and the normal distribution
}
\description{
Demonstrates the Central Limit Theorem by generating the distribution of sums of uniform random variables and
comparing it to the normal distribution.
}
\examples{
# Basic usage with default parameters
myclt(10,1000)

# Custom usage with distribution from -1 to 1
myclt(10,1000,-1,1)

}
