% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{Maximum Likelihood Contour Plot for Normal Distribution}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{A numeric vector containing the sample data}

\item{mu}{A numeric vector of possible mean values to evaluate}

\item{sig}{A numeric vector of possible standard deviation values to evaluate}

\item{...}{Additional graphical parameters passed to the contour function}
}
\value{
A list containing:
  \item{x}{The input data vector}
  \item{coord}{A matrix with row and column indices of the maximum likelihood}
  \item{maxl}{The maximum likelihood value}
}
\description{
Creates a contour plot of the likelihood surface for normal distribution parameters
(mean and standard deviation) based on a sample vector.
}
\examples{
# Generate sample data from a normal distribution
x <- rnorm(50, mean = 5, sd = 2)

# Create contour plot for a range of mean and sd values
result <- mymlnorm(x,
                  mu = seq(4, 6, length = 100),
                  sig = seq(1.5, 2.5, length = 100))

# Compare the MLE coordinates with theoretical values
c(mean(x), sd(x)*(length(x)-1)/length(x))
}
