% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{My Normal Curve Function}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The mean of the normal distribution}

\item{sigma}{The standard deviation of the normal distribution}

\item{a}{The value to shade the area under the curve to the left of}
}
\value{
A list with the mean and standard deviation
}
\description{
Plots a normal probability density function with specified mean and standard deviation
}
\examples{
# Plot a standard normal curve
myncurve(0, 1, 2)

# Plot a normal curve with mean 10 and standard deviation 2
myncurve(10, 2, 12)

}
