% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{Sample Simulation Function}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{Size of sample}

\item{iter}{Number of iterations}

\item{time}{Time between iterations}
}
\value{
A barplot of the sample
}
\description{
Sample Simulation Function
}
\examples{
# Example 1: Basic usage with default parameters
mysample(n = 100)

# Example 2: Increase the number of iterations
mysample(n = 100, iter = 20)

# Example 3: Reduce the sleep time between iterations
mysample(n = 100, iter = 5, time = 0.1)
}
