% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myssr.R
\name{myssr}
\alias{myssr}
\title{Calculate Sum of Squares Residuals and Standard Error}
\usage{
myssr(x, y)
}
\arguments{
\item{x}{Numeric vector containing predictor variable values}

\item{y}{Numeric vector containing response variable values}
}
\value{
A list with two components:
  \item{ssr}{Sum of squares residuals}
  \item{sq}{Standard error of the estimate}
}
\description{
This function performs simple linear regression calculations and returns
the sum of squares residuals (SSR) and standard error of the estimate.
}
\examples{
x <- c(1, 2, 3, 4, 5)
y <- c(2, 3, 5, 4, 6)
myssr(x, y)
}
