% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{pareto}
\alias{pareto}
\title{Pareto Chart}
\usage{
pareto(x, mn = "Pareto barplot", ...)
}
\arguments{
\item{x}{A vector of categorical data for which the Pareto chart will be generated.}

\item{mn}{A character string for the title of the Pareto chart (default is "Pareto barplot").}

\item{...}{Additional arguments passed to barplot}
}
\value{
A Pareto chart (bar plot with a cumulative percentage line).
}
\description{
Pareto Chart
}
\examples{
pareto(c("Bug", "Bug", "Feature Request", "Performance", "Bug"))
}
