% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{getmode}
\alias{getmode}
\title{Mode}
\usage{
getmode(v)
}
\arguments{
\item{v}{A set of data for which the mode should be calculated}
}
\value{
Mode of the data
}
\description{
This function calculates the mode of a set of data
}
\examples{
sample_data <- c(1, 2, 2, 3, 3, 3, 4, 4)
getmode(sample_data)
}
